# Optimise performance of components loading