Lighthouse tested results:

|              | Overall  | FCP       | LCP      | TBT      | CLS     |    SI       |
| ------- | -------- | ------- | ------- | ------- | ------- | ------- | 
| Before  | 42          |  7.8s     | 15s       | 290ms | 0.021    | 10.3s    | 
| After     | 48          | 4.7s      | 9.8s     | 220ms | 0.019    | 7.0s      |